feat: moved manufacturer out of model folder
