+++ src/database/database.module.ts
@@ -27,7 +27,7 @@ import { ReportsModule } from '../reports/reports.module';
 import { AttachmentsModule } from '../attachments/attachments.module';
 import { ConnectionsModule } from '../connections/connections.module';
 import { AddressModule } from '../address/address.module';
-import { ManufacturerModule } from '../model/manufacturer/manufacturer.module';
+import { ManufacturerModule } from '../manufacturer/manufacturer.module';
 
 @Module({
   imports: [
