+++ src/manufacturer/dto/get-manufacturer.dto.ts
@@ -1,6 +1,6 @@
 import { ApiPropertyOptional } from '@nestjs/swagger';
 import { IsOptional } from 'class-validator';
-import { CursorPaginationDto } from '../../../common/dto/cursor-pagination.dto';
+import { CursorPaginationDto } from '../../common/dto/cursor-pagination.dto';
 
 export class GetManufacturersDto extends CursorPaginationDto {
   @IsOptional()
