+++ src/manufacturer/manufacturer.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Manufacturer } from '../model.entity';
+import { Manufacturer } from '../model/model.entity';
 import { ManufacturerService } from './manufacturer.service';
 import { ManufacturerController } from './manufacturer.controller';
-import { CommonModule } from '../../common/common.module';
+import { CommonModule } from '../common/common.module';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Manufacturer]), CommonModule],
